requirements.txt
boto3
Pillow
textract-trp
docx2pdf
PyPDF2
pdf2image
